+++ src/input.js
@@ -2,9 +2,9 @@ const { MODES } = require('./consts');
 
 const { DEDUP_AFTER_LOAD, DEDUP_AS_LOADING } = MODES;
 
-module.exports.validateInput = ({ datasetIds, output, mode, outputTo, preDedupTransformFunction, postDedupTransformFunction}) => {
-    if (!(Array.isArray(datasetIds) && datasetIds.length > 0)) {
-        throw new Error('WRONG INPUT --- Missing datasetIds!');
+module.exports.validateInput = ({ datasetIds, actorOrTaskId, output, mode, outputTo, preDedupTransformFunction, postDedupTransformFunction}) => {
+    if (!(Array.isArray(datasetIds) && datasetIds.length > 0) && !actorOrTaskId) {
+        throw new Error('WRONG INPUT --- Missing datasetIds or actorIdOrName!');
     }
 
     if (!['unique-items', 'duplicate-items', 'nothing'].includes(output)) {